Updates Libraries to Include FTCLib and EasyOpenCV
To give us the most capability when programming our robot, this commit
adds references to the jcenter() repository and pulls libraries for
EasyOpenCV and FTCLib. To be able to further use EasyOpenCv, the
libNative.so library needs to be pushed onto the mobile phones, so that
will be done later.

Adds EasyOpenCV Examples

Since we are going to use EasyOpenCV, the examples have now been added to the project for reference and use.